Move reportError to errors.ts
