+++ src/devTools/editor/fields/ConfigErrorBoundary.tsx
@@ -16,7 +16,7 @@
  */
 
 import React, { Component } from "react";
-import { reportError } from "@/telemetry/rollbar";
+import { reportError } from "@/errors";
 import { getErrorMessage } from "@/errors";
 import { UnknownObject } from "@/types";
 import { faInfoCircle } from "@fortawesome/free-solid-svg-icons";
