+++ src/devTools/editor/panes/insert/GenericInsertPane.tsx
@@ -28,7 +28,7 @@ import { internalExtensionPointMetaFactory } from "@/devTools/editor/extensionPo
 import { ElementConfig } from "@/devTools/editor/extensionPoints/elementConfig";
 import { reportEvent } from "@/telemetry/events";
 import { useToasts } from "react-toast-notifications";
-import { reportError } from "@/telemetry/rollbar";
+import { reportError } from "@/errors";
 import { getCurrentURL, thisTab } from "@/devTools/utils";
 import styles from "./GenericInsertPane.module.scss";
 import {
