+++ src/errors.ts
@@ -16,9 +16,10 @@
  */
 
 import { MessageContext, SerializedError } from "@/core";
-import { deserializeError, ErrorObject } from "serialize-error";
+import { deserializeError, ErrorObject, serializeError } from "serialize-error";
 import { AxiosError } from "axios";
 import { isObject } from "@/utils";
+import { recordError } from "@/background/messenger/api";
 
 const DEFAULT_ERROR_MESSAGE = "Unknown error";
 
@@ -346,4 +347,37 @@ export function selectError(originalError: unknown): Error {
   // Wrap error if an unknown primitive or object
   // e.g. `throw 'Error message'`, which should never be written
   return new Error(String(error));
+}
+
+/**
+ * Report an error for local logs, remote telemetry, etc.
+ * @param error the error object
+ * @param context optional context for error telemetry
+ */
+
+export function reportError(error: unknown, context?: MessageContext): void {
+  void _reportError(error, context).catch((reportingError) => {
+    console.error("An error occurred when reporting an error", {
+      originalError: error,
+      reportingError,
+    });
+  });
+}
+
+// Extracted async function to avoid turning `reportError` into an async function
+// which would trigger `eslint/no-floating-promises` at every `reportError` call
+export async function _reportError(
+  error: unknown, // It might also be an ErrorEvent
+  context?: MessageContext
+): Promise<void> {
+  const errorObject = selectError(error);
+
+  // Events are already natively logged by the browser
+  if (
+    !(error instanceof ErrorEvent || error instanceof PromiseRejectionEvent)
+  ) {
+    console.error(error);
+  }
+
+  recordError(serializeError(errorObject), context, null);
 }
