+++ src/popups/PermissionsPopup.tsx
@@ -21,7 +21,7 @@ import { faShieldAlt } from "@fortawesome/free-solid-svg-icons";
 import AsyncButton from "@/components/AsyncButton";
 import { Button } from "react-bootstrap";
 import Centered from "@/devTools/editor/components/Centered";
-import { reportError } from "@/telemetry/rollbar";
+import { reportError } from "@/errors";
 import browser, { Permissions } from "webextension-polyfill";
 import { getErrorMessage } from "@/errors";
 import { selectOptionalPermissions } from "@/utils/permissions";
