+++ src/telemetry/reportUncaughtErrors.ts
@@ -19,8 +19,7 @@
  * @file This file must be imported as early as possible in each entrypoint, once
  */
 import { isBackground } from "webext-detect-page";
-import { getErrorMessage, IGNORED_ERRORS } from "@/errors";
-import { reportError } from "@/telemetry/rollbar";
+import { getErrorMessage, IGNORED_ERRORS, reportError } from "@/errors";
 
 function ignoreSomeErrors(
   errorEvent: ErrorEvent | PromiseRejectionEvent
